Merged PR 839971: Win 10 UWP SDK sporadic crash in HTTP stack
Win 10 UWP SDK sporadic crash in HTTP stack.

Essentially it's the same fix that was previously done on Win32 Desktop: when we pass the buffer ptr to debug callback, the request has been already released and this may lead to trying to access the heap block possibly re-allocated for something else. The fix is to avoid de-allocation of HTTP req/resp until the handling is done, so when the wrapper is destroyed - the request object is also destroyed.

Related work items: #1510118