Merged PR 977282: Aria SDK changes for DDV Support [Preview]
This PR previews the following change to the Aria SDK:
- Enable a Producer control its own opt-in.
Events fired under the Microsoft Partners Telemetry group (MSPG) are not bound to OS opt-in policies. By permitting the Producer to set its provider group Id, a producer may set its provider group Id to that of MSPG (which UTC maps to a corresponding provider group) and hence control its opt-in.

- Additionally, the addition of new DDV metadata fields on a per-event basis are demonstrated in the sample app included. Discussions on how best to do this are welcome.

Related work items: #1699002