Rename Android classes and artifacts.
Java classes are in the com.microsoft.applications.events package.
The .aar is maesdk-{config}.aar, where {config} is either debug
or release. The aar contains both the Java classes and the eight
flavors of libmaesdk.so (debug/release x 4 ABI flavors).

The Docker image in docker/android-base-image will build the
maesdk aar files. The GitHub runner is unable to complete
a docker build with this image (it runs out of disk space).

The GitHub workflow build-android.yml will build maesdk.