Use sdkmanager to retrieve cmake
The Android gradle plugin insists on specifying the version of cmake
to be used in compiling native code. We can get the specified, Android
supported version of cmake from sdkmanager, and put it in our path to
ensure a stable build (more stable than using whatever version of cmake
we get from choco).