Fix issues in MemoryStorage, OfflineStorage_SQLite, and HttpClient_Android.
MemoryStorage reserved the record rejected by the consumer functor in
GetAndReserveRecords.

OfflineStorage_SQLite destroyed and recreated the database when
it attempted to pass large numbers of IDs into a statement (because
SQLite threw an error).

HttpClient_Android had a race condition that resulted in double-
notification during shutdown (and this in turn caused an assert
in HttpClientManager).