Suppress a warning about deprecated API usage in iOS wrappers for telemetry_ariasdk
Clang now complains because Apple marked this API as deprecated in iOS 8 (not sure when the deprecation tag was added). We cannot yet use the suggested replacement because the new APIs don't exist in Xcode 11. So, write-but-don't-use the code, and suppress the deprecated-api warning in tight scope.

error: 'method:withParam:' is deprecated: first deprecated in iOS 8.0 [-Werror,-Wdeprecated-declarations]