Leak Logger instances to avoid use-after-free crashes
Resolves #481

Each Logger instance has a mutex, a condition variable, and an active-thread counter. When someone calls a method, we increment the active-thread counter, and when we return, we decrement it. When the count hits zero, we wake any condition-variable listeners.

On shutdown, we set our "am I alive?" flag to "Nope." Once the active-thread-count drains to zero, it is safe (as far as this logger is concerned) to destroy the LogManager instance.