Mitigate OOM errors on Android
We are seeing OOM crashes on the oldest and smallest (DRAM) devices.
If we OOM in the database or upload code, we should simply do nothing. We
may succeed on a subsequent attempt, but in any event we do not want to
crash the application because telemetry failed to upload.